LayoutTests/http/tests/security/srcdoc-inherits-referrer.html

<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.dumpChildFramesAsText();
}
</script>
The srcdoc iframe below should use this document's URL as its referrer.
<iframe srcdoc="
This iframe logs information about its referrer:<br>
<iframe src='resources/referrer-policy-log.php'></iframe>"></iframe>
