LayoutTests/http/tests/security/srcdoc-inherits-referrer-expected.txt

The srcdoc iframe below should use this document's URL as its referrer.

--------
Frame: '<!--frame1-->'
--------
This iframe logs information about its referrer:


--------
Frame: '<!--frame2-->'
--------
HTTP Referer header is http://127.0.0.1:8000/security/srcdoc-inherits-referrer.html
Referrer is http://127.0.0.1:8000/security/srcdoc-inherits-referrer.html
